error fixes

+++ src/controllers/regimeControllers.js
@@ -10,13 +10,15 @@ export async function getAllRegimePlans(req, res) {
 }
 export async function addRegimePlan(req, res) {
   const { name, description } = req.body;
+  const regimeImage = req.file.path
   const newPlan = new Regime({
     name: name,
     description: description,
+    image:regimeImage
   });
   try {
     await newPlan.save();
-    res.json({ message: "New plan have been created" });
+    res.json({ message: "New plan have been created", data: newPlan });
   } catch (error) {
     console.log(error);
   }
@@ -29,11 +31,11 @@ export async function addRegimePlan(req, res) {
       const regimeImage = req.file?.path || target.image;
   
       try {
-        const data = await Regime.findOneAndUpdate(
+        await Regime.updateOne(
           { _id: id },
           { name: name, description: description, image: regimeImage }
         );
-        res.json({ data: data });
+        res.json({ message:"Updated Successfuly" });
       } catch (error) {
         console.log(error);
       }
@@ -44,10 +46,16 @@ export async function addRegimePlan(req, res) {
   
 export async function removeRegimePlan(req,res){
     const id = req.body.id
-    try {
-        await Regime.findOneAndDelete({_id: id})
-        res.json({message: 'Deleted Successfuly'})
-    } catch (error) {
-            console.log(error)
+    if(id){
+
+      try {
+        const regimePlan = await Regime.findOneAndDelete({_id: id})
+        regimePlan? res.json({message: 'Deleted Successfuly'}) : res.json({message: `No regime plan with the id ${id}`})
+      } catch (error) {
+        console.log(error)
+
+      }
+    }else{
+      res.json({message: "Provide an id"})
     }
 }
